Fix Float Contrastable.Foundation given mistyped as Byte
`given float` declared `Byte is Contrastable.Foundation`, duplicating
`byte` — so resolving `Byte is Contrastable.Foundation` was ambiguous
(and there was no Float instance at all). The ambiguity made the
`Contrastable.derived` summonFrom silently fail for `Byte`, surfacing
only now that contrast expectations are enabled by default. Correct the
type to `Float` and widen through `double` rather than truncating to
`long`.

+++ lib/chiaroscuro/src/core/chiaroscuro.Contrastable.scala
@@ -91,7 +91,7 @@ object Contrastable:
     given int: Int is Contrastable.Foundation = long.juxtaposition(_, _)
     given short: Short is Contrastable.Foundation = long.juxtaposition(_, _)
     given byte: Byte is Contrastable.Foundation = long.juxtaposition(_, _)
-    given float: Byte is Contrastable.Foundation = long.juxtaposition(_, _)
+    given float: Float is Contrastable.Foundation = double.juxtaposition(_, _)
 
     given set: [element: Showable] => Set[element] is Contrastable.Foundation =
       (left, right) =>
